# server/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1.0",
    "title": "Equipment Distribution APIs",
    "description": "API for Equipment Distibution System built by <b>Bruce</b> using NodeJS and Mysql."
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "Authorization": {
      "type": "apiKey",
      "name": "Authorization",
      "description": "Value: Bearer ",
      "in": "header",
      "scheme": "bearer"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginModel"
            }
          }
        ],
        "responses": {
          "200":{
            "description":"Login successfully"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RegisterModel"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "409": {
            "description": "User already Exist"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "definitions": {
    "LoginModel": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string",
          "example": "rice"
        },
        "password": {
          "type": "string",
          "example": "rice12"
        }
      },
      "required": [
        "username",
        "password"
      ]
    },
    "RegisterModel": {
      "type": "object",
      "properties": {
        "names": {
          "type": "string",
          "example": "John"
        },
        "username": {
          "type": "string",
          "example": "Snow"
        },
        "password": {
          "type": "string",
          "example": "nezabruce@gmail.com"
        }
      },
      "required": [
        "names",
        "username",
        "password"
      ]
    },
...
  }
}